Implemented mat_mul_trans test, fixed mat_mul test (no @ operator)
