+++ test/mrWolf/mat_mul/test_lib/riscy/gen_stimuli.py
@@ -23,8 +23,7 @@ def compute_result(result_parameter, inputs, env, fix_point):
         a = inputs['srcA'].value.astype(np.int32).reshape((env['len_m'], env['len_n']))
         b = inputs['srcB'].value.astype(np.int32).reshape((env['len_n'], env['len_o']))
         if fix_point is None or fix_point == 0:
-            result = (a @ b).astype(np.int32)
-            result = result.reshape((env['len_res'], ))
+            result = np.matmul(a, b).astype(np.int32).reshape((env['len_res'], ))
         else:
             raise RuntimeError("Fix-Point not implemented")
     elif result_parameter.ctype == 'float':
