+++ test/mrWolf/mat_mul_trans/test_lib/riscy/cluster.h
@@ -1,0 +1,6 @@
+#ifndef __INC_CLUSTER_H__
+#define __INC_CLUSTER_H__
+
+void cluster_entry(void *arg);
+
+#endif
