+++ test/mrWolf/mat_mul_trans/test_lib/testset.cfg
@@ -1,0 +1,19 @@
+from plptest import * 
+
+TestConfig = c = {}
+
+ibex_tests = Testset(
+  name  = 'ibex',
+  files = [
+    'ibex/testset.cfg',
+  ]
+)
+
+riscy_tests = Testset(
+  name  = 'riscy',
+  files = [
+    'riscy/testset.cfg',
+  ]
+)
+
+c['testsets'] = [ ibex_tests, riscy_tests ]
