+++ test/mrWolf/testset.cfg
@@ -18,3 +18,4 @@ add_test_folder(c, 'conv_valid')
 add_test_folder(c, 'conv_valid_rep')
 add_test_folder(c, 'dot_prod')
 add_test_folder(c, 'mat_mul')
+add_test_folder(c, 'mat_mul_trans')
